+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Master
+##### Enhancements
+* Open template from URL String [#8](https://github.com/toshi0383/TVMLKitchen/pull/8)
+
 ## 0.1.2
 ##### Breaking Change
 * Improve error handling: [#4](https://github.com/toshi0383/TVMLKitchen/issues/4)